User guesses a character & game logic if game is won.
Made it so that there is are dashes in places of the characters of the
word to be guessed. If the character is guessed, it will replace the
dash with the correct character. Allows player to guess the word after
each character guess attempt. Added a way to distribute points based on
difficulty on the default game mode. Only added logic for if
word/character is guessed correctly.