Added to show incorrect guesses
Also made it not count previously guessed characters.
changed the game so that the user can guess the word only after he has
failed to guess all of the characters(completed the hanged man) changed
the guess to be a char that is at index 0 of whatever the user inputed.
added a stack to display every incorrect guesses in the order that they
were guessed.